requirements.txt
flask
paho-mqtt
network
requests
time
machine
dht
